Move routes to storage
