+++ aikido_zen/background_process/commands/sync_data_test.py
@@ -1,9 +1,7 @@
-from multiprocessing.forkserver import connect_to_new_process
-
 import pytest
 from unittest.mock import MagicMock
 from .sync_data import process_sync_data
-from aikido_zen.background_process.routes import Routes
+from aikido_zen.storage.routes import Routes
 from aikido_zen.helpers.iplist import IPList
 from ..packages import PackagesStore
 from ...storage.hostnames import Hostnames
